@@ -1,118 +1,117 @@
-﻿namespace Serilog.Sinks.XUnit.Tests
+﻿using AwesomeAssertions;
+
+namespace Serilog.Sinks.XUnit.Tests;
+
+using System;
+using System.IO;
+using Core;
+using Events;
+using Formatting;
+using NSubstitute;
+using Xunit;
+
+public static class TestOutputHelperExtensionsTests
 {
-    using System;
-    using System.IO;
-    using Core;
-    using Events;
-    using FluentAssertions;
-    using Formatting;
-    using NSubstitute;
-    using Xunit;
-    using Xunit.Abstractions;
-
-    public static class TestOutputHelperExtensionsTests
+    [Fact]
+    public static void CreateTestLogger_WithDefaultParameters()
     {
-        [Fact]
-        public static void CreateTestLogger_WithDefaultParameters()
-        {
-            var outputMock = Substitute.For<ITestOutputHelper>();
-            var logger = outputMock.CreateTestLogger();
+        var outputMock = Substitute.For<ITestOutputHelper>();
+        var logger = outputMock.CreateTestLogger();
 
-            const string message = "This is a test message";
-            logger.Information(message);
+        const string message = "This is a test message";
+        logger.Information(message);
 
-            outputMock.Received(1).WriteLine(Arg.Is<string>(text => text.Contains($" INF] {message}")));
-        }
+        outputMock.Received(1).WriteLine(Arg.Is<string>(text => text.Contains($" INF] {message}")));
+    }
 
-        [Fact]
-        public static void CreateTestLogger_WithCustomMessageTemplate()
-        {
-            var outputMock = Substitute.For<ITestOutputHelper>();
-            var logger = outputMock.CreateTestLogger(outputTemplate: "Game of Thrones {Level}: {Message}");
+    [Fact]
+    public static void CreateTestLogger_WithCustomMessageTemplate()
+    {
+        var outputMock = Substitute.For<ITestOutputHelper>();
+        var logger = outputMock.CreateTestLogger(outputTemplate: "Game of Thrones {Level}: {Message}");
 
-            const string message = "That's what I do. I drink and I know things.";
-            logger.Warning(message);
+        const string message = "That's what I do. I drink and I know things.";
+        logger.Warning(message);
 
-            outputMock.Received(1).WriteLine(Arg.Is<string>(text => text.Equals($"Game of Thrones Warning: {message}")));
-        }
+        outputMock.Received(1).WriteLine(Arg.Is<string>(text => text.Equals($"Game of Thrones Warning: {message}")));
+    }
 
-        [Fact]
-        public static void CreateTestLogger_WithCustomMessageTemplateWithExtraNewline()
-        {
-            var outputMock = Substitute.For<ITestOutputHelper>();
-            var logger = outputMock.CreateTestLogger(outputTemplate: "Game of Thrones {Level}: {Message}{NewLine}{Exception}");
+    [Fact]
+    public static void CreateTestLogger_WithCustomMessageTemplateWithExtraNewline()
+    {
+        var outputMock = Substitute.For<ITestOutputHelper>();
+        var logger = outputMock.CreateTestLogger(outputTemplate: "Game of Thrones {Level}: {Message}{NewLine}{Exception}");
 
-            const string message = "That's what I do. I drink and I know things.";
-            logger.Warning(message);
+        const string message = "That's what I do. I drink and I know things.";
+        logger.Warning(message);
 
-            outputMock.Received(1).WriteLine(Arg.Is<string>(text => text.Equals($"Game of Thrones Warning: {message}")));
-        }
+        outputMock.Received(1).WriteLine(Arg.Is<string>(text => text.Equals($"Game of Thrones Warning: {message}")));
+    }
 
-        [Fact]
-        public static void CreateTestLogger_WithCustomMinimumLogLevel()
-        {
-            var outputMock = Substitute.For<ITestOutputHelper>();
-            var logger = outputMock.CreateTestLogger(LogEventLevel.Error);
+    [Fact]
+    public static void CreateTestLogger_WithCustomMinimumLogLevel()
+    {
+        var outputMock = Substitute.For<ITestOutputHelper>();
+        var logger = outputMock.CreateTestLogger(LogEventLevel.Error);
 
-            const string message = "Foo";
-            logger.Debug(message);
-            logger.Information(message);
-            logger.Warning(message);
+        const string message = "Foo";
+        logger.Debug(message);
+        logger.Information(message);
+        logger.Warning(message);
 
-            outputMock.DidNotReceive().WriteLine(Arg.Any<string>());
-        }
+        outputMock.DidNotReceive().WriteLine(Arg.Any<string>());
+    }
 
-        [Fact]
-        public static void CreateTestLogger_WithCustomSwitch()
-        {
-            var outputMock = Substitute.For<ITestOutputHelper>();
-            var levelSwitch = new LoggingLevelSwitch(LogEventLevel.Warning);
-            var logger = outputMock.CreateTestLogger(levelSwitch: levelSwitch);
+    [Fact]
+    public static void CreateTestLogger_WithCustomSwitch()
+    {
+        var outputMock = Substitute.For<ITestOutputHelper>();
+        var levelSwitch = new LoggingLevelSwitch(LogEventLevel.Warning);
+        var logger = outputMock.CreateTestLogger(levelSwitch: levelSwitch);
 
-            const string message = "Bar";
-            logger.Verbose(message);
-            logger.Debug(message);
-            logger.Information(message);
-            outputMock.DidNotReceive().WriteLine(Arg.Any<string>());
+        const string message = "Bar";
+        logger.Verbose(message);
+        logger.Debug(message);
+        logger.Information(message);
+        outputMock.DidNotReceive().WriteLine(Arg.Any<string>());
 
-            levelSwitch.MinimumLevel = LogEventLevel.Information;
+        levelSwitch.MinimumLevel = LogEventLevel.Information;
 
-            logger.Information(message);
+        logger.Information(message);
 
-            outputMock.Received(1).WriteLine(Arg.Is<string>(text => text.Contains(message)));
-        }
+        outputMock.Received(1).WriteLine(Arg.Is<string>(text => text.Contains(message)));
+    }
 
-        [Fact]
-        public static void CreateTestLogger_WithCustomTextFormatter()
-        {
-            var outputMock = Substitute.For<ITestOutputHelper>();
-            var textFormatter = Substitute.For<ITextFormatter>();
-            var logger = outputMock.CreateTestLogger(textFormatter);
+    [Fact]
+    public static void CreateTestLogger_WithCustomTextFormatter()
+    {
+        var outputMock = Substitute.For<ITestOutputHelper>();
+        var textFormatter = Substitute.For<ITextFormatter>();
+        var logger = outputMock.CreateTestLogger(textFormatter);
 
-            const string message = "Baz";
-            logger.Error(message);
+        const string message = "Baz";
+        logger.Error(message);
 
-            textFormatter.Received(1).Format(
-                Arg.Is<LogEvent>(logEvent => logEvent.Level == LogEventLevel.Error && logEvent.MessageTemplate.Text == message),
-                Arg.Any<StringWriter>());
-        }
+        textFormatter.Received(1).Format(
+            Arg.Is<LogEvent>(logEvent => logEvent.Level == LogEventLevel.Error && logEvent.MessageTemplate.Text == message),
+            Arg.Any<StringWriter>());
+    }
 
-        [Fact]
-        public static void CreateTestLogger_ShouldThrowIfTestOutputHelperIsNull()
-        {
-            Action act = () => default(ITestOutputHelper).CreateTestLogger();
+    [Fact]
+    public static void CreateTestLogger_ShouldThrowIfTestOutputHelperIsNull()
+    {
+        Action act = () => default(ITestOutputHelper)!.CreateTestLogger();
 
-            act.Should().Throw<ArgumentNullException>()
-                .And.ParamName.Should().Be("testOutputHelper");
-        }
+        act.Should().Throw<ArgumentNullException>()
+            .And.ParamName.Should().Be("testOutputHelper");
+    }
 
-        [Fact]
-        public static void CreateTestLogger_ShouldThrowIfFormatterIsNull()
-        {
-            Action act = () => Substitute.For<ITestOutputHelper>().CreateTestLogger(default(ITextFormatter));
+    [Fact]
+    public static void CreateTestLogger_ShouldThrowIfFormatterIsNull()
+    {
+        Action act = () => Substitute.For<ITestOutputHelper>().CreateTestLogger(default(ITextFormatter)!);
 
-            act.Should().Throw<ArgumentNullException>()
-                .And.ParamName.Should().Be("formatter");
-        }
+        act.Should().Throw<ArgumentNullException>()
+            .And.ParamName.Should().Be("formatter");
     }
 }
